// src/app/interfaces/lead/lead.ts
export interface Lead {
  zip: string;
  leadId: string;
  firstName: string;
  lastName: string;
  title: string;
  department: string;
  accountName: string;
  salesRep: string;
  email: string;
  leadSource: string;
  phone: string;
  status: string;
  coporateIdentity: string;
  description: string;
  referredBy: string;
  billingAddress1: string;
  billingAddress2: string;
  billingCity: string;
  billingState: string;
  billingZip: string;
  shippingAddress1: string;
  shippingAddress2: string;
  shippingCity: string;
  shippingState: string;
  shippingZip: string;
}